Split the Micromega library in two components.
The first component is the part of the library used by csdpcert, which only
relies on basic APIs from clib. The second one embarks all the Ltac-related
code and corresponds to the previous micromega plugin.

Since csdpcert only links the first one, its executable doesn't link findlib
anymore, which prevents implicitly setting the -linkall flag. This was useless
for csdpcert since this executable never performs any plugin dynlink. This
reduces the size of the csdpcert binary from 32Mo to 6.5Mo on my machine.

Maybe the barebone library should live in its own subdirectory but it was
very annoying w.r.t. the other parts of Coq's build system so I left it there.