Put parts of ltac2_plugin which depend on ltac1 in a separate plugin
This requires a new `ltacX_common_plugin` for things used by both
(profile_ltac, goal selectors and the "use_default" `...` syntax).
We could also put that stuff in vernac/.

For now we still load the split off parts in Ltac2.Init.

It could also make sense to load it in Ltac2.Ltac1 (and require that
in Ltac2.Notations as a few notations use `ltac1:()`).