Fix #18281 (plan B).
When converting a primitive projection with the corresponding opaque
compabitility constant, we force the unfolding of the compatibility
constant, and also make the projection unfolded.

Both steps are required, since compabitiliy constants are defined to
be unfolded projections, and unfolded projectons are pushed to the
stack during weak-head reduction, while folded projections remain in
head position.