Make light CI lighter
The vio jobs don't have a reason to be in light CI IMO.

The validate jobs are largely redundant with the coqchk runs in the
test suite.

We don't need multiple variants of the test suite.

We don't need mutliple variants of the build job (the nix job can be
seen as a variant of the build job).
(lint uses edge ocaml, so have the base build + the lint job is enough
to check we don't use too new or too old ocaml APIs)