[stm] Remove support for `.vio` files
As discussed in https://github.com/coq/coq/wiki/Coq-Call-2023-07-11 ,
we do this PR as a first step to evaluate what are the gains when
removing `.vio` files.

Due to #4123, it is not likely that `.vio` files are useful in
practice, and `.vos` files provide a viable alternative (as outlined
in the original vos PR)

Moreover, we have no plans to address the remaining vio bugs in the
issue tracker.

This PR:

- closes #4007
- closes #4013
- closes #4123
- closes #5223
- closes #5308
- closes #6720
- closes #8402
- closes #9637
- closes #11471
- closes #18380

Related: #5388 #6710

See also the previous attempt: #8767 #8783