Reconstruction of the intent of the use of history in Rtree.inter.
For efficiency, Rtree.inter tries to preserve as much of the common
structure of the two trees to intersect, so expansions are done
lazily.

Rtree.inter uses an history and I understand that this history is used
to ensure that only a finite number of expansions is needed to manage
overlapping expansions.

My understanding is that the effective expansion of variables is
missing when using this history: this is a reason for an
incompletenesses of Rtree.inter, which sometimes falls and fails in an
unsupported Rec/Var case.

The commit is fixing this, by instantiating the variables before using
the history.

A typical example of incompleteness before the fix is when comparing
the following equivalent trees:

  Rec{Mrec[Wrapper, 0],
      [(Rec{Nested[Unwrapper, 0], [(), (#1:0)]})]}

and

  Rec{Nested[Wrapper, 0],
      [(Rec{Mrec[Unwrapper, 0], [(), (Rec{Nested[Wrapper, 0], [(#1:0)]})]})]}

(example from success/fix.v)