Design choice: hide as most implicit arguments as possible in notations.
That is, use the longest set of implicit arguments when several blocks
of implicit arguments are available.

In practice, this gives a nicer display when maximal implicit
arguments are not explicitly part of the body of the
notation. Currently, this is the case only with atomic constant, but
when done also for applied constants, and one gets a better
compatibility by taking the longest block. Actually, this is also
already the case when not using notations (e.g. "Check (eq_refl 1)"
prints "eq_refl").