Algorithmically efficient evar expansion.
We merge identifier substitution and recursive evar expansion in one single
pass. The algorithm is a bit unsatisfactory as it relies on a cache to ensure
we neither have to recompute an offset in time linear in the depth of evar
nesting when substituting identifiers, nor compute upfront these offsets when
reaching a defined evar. But in the end, this should now be basically linear
it the size of the term (modulo some logarithmic overheads).