Move coqperf to a separate opam package
This makes it optional on linux instead of forcing everyone to install
kernel headers.

We could also move it to a plugin (using a hook in lib I guess)
so that installing it doesn't require recompiling everything. Note
that this would mean instruction counts in NewProfile would only be
available after the plugin is loaded though.