Ensure transparency of a constant of same name as a transparent primitive projection.
We do it at the exact point where it matters.

Note: for constants associated to a primitive projection not set
transparent in a "delta [constants]" expression, the constant is kept
opaque (in a model where these constants are eta-expansion of the
primitive projection, it is somehow "beta" that we would need to
reduce them).