Ltac2 #[abstract] attribute for types
Fix #18656

Printing of glb_expr is a bit awkward. We don't have access to the
type so we don't know when we cross an abstraction barrier. OTOH
constructors and projections are not in the environment / nametab so
we still print <abstr> in some places.

The attribute is disabled for open types as the naive implementation
produces not necessarily expected results.