Always set `Sys.catch_break true`, not just when interactive
When debugging it can be nice to run coqc and try to get a backtrace
from interrupting it.

eg https://coq.zulipchat.com/#narrow/stream/237656-Coq-devs-.26-plugin-devs/topic/Debugging.20hangs

OTOH if we're hanging in such a way that we can't process signals this
may be a bad idea so maybe we should only enable it with backtraces in
non interactive mode?