Rely on evar rather than metas in canonical structure resolution.
I am not sure how critical this is for unification heuristics, e.g. when
considering HO vs FO algorithms. This piece of code was problematic because
before this patch it was the only one in the whole code base that created
dangling metas not well-scoped within a clausenv. The alternative is to
try to prune these freshly generated metas after resolution has run, but this
is probably too dangerous and unsound in practice.