Put evar counter in the evar map instead of summary
This should make it less a footgun when resetting summaries such as in
https://coq.zulipchat.com/#narrow/stream/237656-Coq-devs-.26-plugin-devs/topic/When.20does.20.60new_type_evar.60.20return.20an.20instantiated.20evar.3F

We could perhaps do the same for the meta counter but not sure if
worth doing.

NB: the original use of summary was in
81429f879fa76ac24c25d89a6ada2073ea8361e2

>Registering universe and meta/evar counters as summaries so as to
eventually get the same numbers when replaying

Because universes escape `par:` and vio they ended up having to be
treated specially but this should not be a problem for evars.