Mark arrays, products, and plain constants, as potentially needing a delayed evaluation during native_compute (fix #18703).
The following statements would otherwise cause a stack overflow if the
module were to be compiled and then loaded, even if the constant was
completely unrelated to the term being normalized.

  Definition large := [|fact 100|0|].
  Definition large := (0 <= fact 100) -> Prop.
  Definition large := large'.

Note that the following statements are still eagerly evaluated, on purpose.

  Definition small := [|0|0|].
  Definition small := Prop -> Prop.

Note also that the translation of the following code is a bit suboptimal:

  Definition large := large'.
  (* let large = lazy (Lazy.force large') *)