Prevent side effect temporary global env entries from leaking
Fix #13324
Also fix #13271

The bug occurs in 2 code paths so there are 2 changes
- Proof.refine_by_tactic for ltac:() / ltac2:() / whatever rewrite.ml is doing
- Declare.solve_by_tac sor Solve Obligations / automatic obligation solving.

We might have wrapped Declare.build_by_tactic in state protection
everywhere instead of just the solve_by_tac call, but this makes
funind fail. Not sure if funind could be changed to make it work.

There is also a use of inline_private_constants in
obligation_terminator, but the state reset is handled by the stm for
this AFAICT.