Use gitlab.inria.fr shared runners for docker boot
Thanks to Guillaume Melquiond for telling us this is possible.

Behind the scenes to avoid using the ci.inria.fr workers for other
jobs I retagged the ci-coq workers adding the `ci-coq` tag
and modified variable TAGGED_RUNNERS := ci-coq in the project
settings.

Once the pipelines started before this change have finished we can
enable the shared runners and test this commit.

As mentioned in the comments making the tags for the docker job
configurable by variable seems too annoying to be worth doing so I
hardcoded them.