More efficient implementation of UState.demote_global_univs.
The only user out there is coq-elpi, which relies on it to work around a
deficient API for global declaration. Intuitively, they need to remove
the globally declared universes from an evarmap they want to reuse.

The previous implementation was extremely naive, rebuilding the whole set
of global constraints from stratch. As a result this function accounted for
about 1/3 of the runtime of elpi programs. We replace this by a still naive
but more efficient code. After this commit this hotspot seems to have simply
vanished.