Cleanup template poly syntax restriction handling
No need to pass an optional pointer to the inferred sort (which is
incorrect when lowering to Prop) around, we just need a boolean to
tell if the syntax was the kind we allow for template poly.

This allows to get rid of special handling of prop lowering in the
auto template branch of compute_template_inductive.

This changes behaviour for records:
~~~coq
Polymorphic Definition typ := Type.

Record foo (A:Type) : typ := { x : A }.
~~~
would be made template poly and now is not.