Set some basic rewriting primitives in Type opaque for TC search.
The corresponding Prop primitives are already set opaque in Classes.Init,
hence I think this is an overview. These low-level primitives being transparent
mean that TC search when setoid rewriting in Type fires off a crapload of
search branches that should not be there to begin with. Hopefully nobody relies
on that in the wild.