Make MAKESWITCHBLOCK more robust against large stack frames.
When it encounters an accumulator, MAKESWITCHBLOCK copies the whole stack
frame of the current function, so that the code of the pattern-matching
branches can be replayed later, during strong normalization. There is no
reason why the stack frame could not exceed 256 words, even in a
non-malicious function, so it needs to be taken into account.