Fast path in Unification.make_abstraction_core.
We keep track of the variables available from the context in order to
skip declarations when the pattern mentions variables not yet introduced.
This is very slightly non-equivalent to the previous code as a variable
may disappear by reduction through unification. But the code already makes
a similar assumption when computing dependent declarations. In practice it
should not matter.