Stricter criterion for template universe levels.
We restrict template levels so that we can substitute them with arbitrary
algebraic levels. The criterion is an overapproximation, and should be
removed when algebraic universes are made first-class in the kernel.

The upper layers are actually already implementing this check, so that the
kernel could not have received ill-formed template inductive types from
user input alone.