Track unused constant arguments in Genlambda and erase them.
We replace arguments that do not appear in the lambda body of constants
by a dummy constant value. The rationale is that most arguments in symbolic
code are type annotations that have no computational content, hence there is
no point in generating them in the first place.

This patch seems to have wonderful effects on degenerate examples from UniMath.