Use the same criterion in the VM and native for substitutable atoms.
We use the criterion from the VM rather than the native one, because it is
more reasonable. The code for the VM allowed additionally 63-bit integers,
while the one for native allowed λ-abstractions and evars. Duplicating
the latter, which may contain arbitrary subterms, does not sound like a
great idea for code size.