Faster composed module substitution of terms.
Rather than folding over the list, which is repeatingly mapping over the whole
term, we map once over the term and perform the substitution of constants there.
Most of the time the term is orders of magnitudes bigger than the list, as this
function is typically used to force the body of opaque proof terms.

This source of slowdown was observed when calling the checker on compcert.