Use irrelevance masks to ignore arguments in flex / flex conversions.
We reuse the constant masks computed by the VM to skip conversion of arguments
that are runtime irrelevant. This is sound, since if these arguments mattered
for conversion the VM compilation should have kept them around. Indeed, VM
conversion is as untyped as kernel conversion, so that both algorithms have
access to the same information.