Factorize the code for unification of universes.
This is a slight change of behaviour, as we are now relying on a FO
heuristic for compatibility of universes. The previous code would unfold
constants when faced with a local inconsistency, but now this problem will
be deferred to after the conversion phase and will result in a unification
failure.