Remove transitive dependencies from vo summaries.
This information can be recomputed, and keeping it results in a quadratic
blowup of the vo summary field. This is more tricky than it sounds because
the precise topological order of require matters, as some superglobal
libobjects are not idempotent. We keep track of the explicit dependencies
to be able to recompute the precise order we obtained after the chain of
requires.

As a result, we could still have a quadratic blowup in the worst case, but
this patch makes it unlikely in the average case.