[library] Make error handling explicit when locating libraries
It is still a bit open when to check for errors, for now this is just
a proposal, which allows internalization functions to raise errors.

It would be good to think what other errors we'd like to capture in
the type, certainly at least `Invalid_file` for wrong .vo files would
be great to have in the type.