[vernac] [experiment] Reorganization of parsing state.
As of today, we keep copies of both parsing state in the synterp
summary and in the record. This is redundant and in fact makes
`Synterp.unfreeze` to perform useless work as it calls `Pcoq.unfreeze`
twice.

We try to remove this duplicity and rework a bit the vernacstate
organization, so parsing state is a regular entry of synterp state.

This is more flexible (and I prefer that as to be ready for the day
`Pcoq.parse st entry pa` could be understood as pure), but we need to
be careful, there are some hidden invariants in the unfreze dance.

Let's see what CI finds about this experiment.