[vernacstate] Remove duplicate copy of the parsing state.
As noted in #19159 (of which this PR is an alternative), `Vernacstate`
stores a copy of the parsing state, which is also already stored in
the summary.

Thus, we do unnecessary job on unfreeze.

We remove this duplication and choose to keep the parsing state in the
summary, which seems to work, and seems OK for now as more explicit
state management is tricky as noted by Gaëtan Gilbert: sections
messing with summaries behind the back of `Vernacstate`, so they would
have to restore `Pcoq.frozen_t` too as to support this approach.

There is no real advantage to a separate `Pcoq` state for now.