Ltac2 don't print empty backtrace
This avoids printing an empty "Backtrace:" entry next to the real one
when Ltac2 Backtrace is on.

(it gets printed when an exn is wrapped, because the wrapping uses
CErrors.print which gives an empty exninfo)

It also avoids printing en empty "Backtrace:" for exception which have
nothing to do with ltac2 (eg `Check Prop : Prop.`).