Faster guardedness check when checking non-recursivity.
Instead of performing an expensive unfolding of the rectree to check that
an argument is Norec, we simply look directly at the head constructor. Indeed,
the only way to get a Node constructor by unfolding is when the argument is
itself a Node, i.e. unfolding is useless when we only care about the head.