Update `flake.lock` to re-enable `nix build`
Currently `nix build` through the flake fails with:
`Error: Version 3.6 of the dune language is not supported.`

I just ran `nix flake update` and checked that the `nix build`
didn't fail. I also added the result symlink to the `.gitignore`
so that running `nix build` doesn't dirty the git repo.