Fix double substitution of relevance annotations in cclosure
Fix #19325

Binders are already documented to be non substituted in cclosure.mli,
this fixes the code to match the documentation.

AFAICT the bug can't be exploited in the kernel in coqc without a
plugin, because the toplevel instance will not contain Var so
substitution by it is idempotent.

It can probably be exploited by a plugin (which can declare constants
pre-univpoly-abstracted) and in coqchk.

(cherry picked from commit b43cb7834e18f79597d3adabfc31d34ce7e9bbcb)