Avoid allocating a new term when already whnf in Reductionops.whd_betaiota
This is crucially used with `strongrec` in `meta_instance` such that
when we unify `S ?M == S (S ...)` without this code the value we get
for ?M is not shared with the RHS at all. eg in

~~~coq
Inductive is_nat : nat -> Prop :=
| Is_nat_Z : is_nat 0
| Is_nat_S : forall n, is_nat n -> is_nat (S n).

Lemma is_nat_500 : is_nat 500.
Proof.
  repeat constructor.
  (* proof = Is_nat_S 499 (Is_nat_S 498 ...) *)
Time Qed.
~~~
all the copies of 1, 2, 3, ..., 499 are not shared with each other
without this change.

(example from #18520)