Faster coherence check when adding coercions.
When the coherence warning is unset, which is the case for heavy user of
coercions, we do not perform an expensive traversal of the graph.
The underlying algorithm is still very bad as it is at least quadratic in the
number of classes, but this patch should alleviate a good part of the issues.