Reduce the Typeops API.
We remove a bunch of functions that were never really used, which included
all the judge_* functions. Indeed, in the current implementation the kernel
typechecker is merely a blackbox used by the kernel alone. The typing
functions from the upper layers are all reimplemented from scratch to handle
the unification state, and as a result the small-step variants of the kernel
typing rules are not needed.

Reducing the exposed API also results in less coupling between what the kernel
does and what the toplevel users rely on. In particular it makes it easier to
tweak the kernel typing rules (e.g. for efficiency) without breaking the
user-facing ones.