Add visibility rules in section in module
A local definition in a section is also local in the current module.
An export/global setting in a section is also export/global in the
current module. Unfortunately, this makes the section table more
complicated.

The section part of the refman is also slightly modified:
- a subtitle "Using sections" appears before the description of the
  commands, in order to have a better TOC
- In the first sentences, "declarations" was inaccurate, since Let
  introduces definitions
- The "Note" starting with "Most commands" refers to the locality table

(cherry picked from commit 7d93822b72c277ca0358be90062cee84ec13ba46)