Small implementation tweak in ring protection.
The previous code was recursively normalizing arguments of the shape App (R, args)
where R is the head relation of the toplevel expression. In practice, this
does not make much sense since no subterm of the expression being normalized
can be a relation. The intention seems to have been instead to normalize
the arguments of the head relation without touching the rest, and it was
implemented this way because it was simply easier to code.

After this commit we simply normalize directly the arguments of the head relation
without performing anything else for this relation in the subterms