Rely on the HConstr machinery also for opaque proofs containing side-effects.
The previous code was performing some surgery with the input term. Although it
used to matter because the kernel could fix relevance marks, this is nowadays
useless. The term return by Typeops is the same as the input one. We leverage
this to return the hashconsed term in more cases, leading to a more efficient
term hashconsing down the road.