Unconditionally use the fresh mechanism in detyping.
We have to pay an upfront cost linear in the size of the context
everytime call detyping, but this should be compensated by the fact
that the name generating algorithm went from quadratic to quasilinear.

Furthermore the cost is only paid when the goal_concl_style flag is
set, which is only the case for the conclusion of a goal so, barring
a few Himsg printers, we already pay a linear cost in this case as
we also print along the named context.