ci/README-users: plugins don't have to compile with Coq's ocaml flags
We don't enforce this and IIRC we have some plugins in CI which don't
respect this. Also since the Coq build switched to dune it is harder
to get the flags used to compile Coq (we still ship some default flags
in coq_makefile but I don't think we keep them synchronized with the
Coq build).