interp/stdarg: add a GADT type of all basic generic types
When someone adds a new basic generic type, it is easy to forget to
register a printer for it. To avoid this, we introduce a sum type of
all basic generic types. Operations that should be applied to each
basic generic type can now be defined by pattern-matching on the sum
type, and adding a new generic type without extending them will result
in an exhaustivity error.