Refresh template levels in clenvs.
When creating a clenv out of a template polymorphic constructor, we
now refresh the type of the arguments. This prevents adding spurious global
constraints between the argument types and the global default template level
when calling Clenv-based tactics such as apply. The implementation is a bit
hackish because some callers fiddle with clenv evarmaps but in practice it
seems to be mostly working.