Fix unicode classification of non-spacing marks.
This commit adds the ~2000 non-spacing marks into the IdentPart category.
This includes all the combining marks, and thus fixes #19512. This also
means that characters in the range 1DC0-1DFF can no longer appear at the
start of an identifier (which does not make sense anyway, as they are
combining marks).

This commit also fixes a few exceptions, which were actually no exception:
- the dot is already in Symbol,
- phonetic extensions are already in Letter.