Rocqify the CoqIDE source code.
Since coqtop has a specific meaning in this context, I also rocqified it when
it was mentioned in the code or in the comments. The user-facing display should
not have been touched by this commit though. Some cases were a bit liminal, so
in case of doubt I didn't modify them.