reachability analysis for hconstr
because of_constr recurses on binder types in matches, some of the
terms it generates are not reachable

this means that eg (with `A` a closed term)
~~~
fun x : option A =>
match x with None => true | Some _ => false end
~~~
sees `A` twice (once in `option A`, then again
as the type of the anonymous argument of `Some`)
and before this commit would give it refcount 2.

Then typeops and hconsing would pointlessly cache their result on `A`.