Use the release profile for configure in opam
The build instructions in the opam files correctly use `-p` for the
builds themselves, but the configure script uses the default profile.
The configure script now recognises -release in its first argument as
meaning that `--profile=release` should be added to the `dune exec`
invocation, and the opam files are updated to include this flag.