Delay import of notation sets in Notationextern.
This speeds up a lot importing files that define a lot of notations. The
underlying algorithm is still very inefficient and quite non-sensical, as
it adds elements in a list in O(n) w.r.t. an order which is not total.
We should reimplement it in a more sensible way but that requires a redesign,
e.g. using some kind of dnets. In the meantime this patch is simple and
seems to solve a good part of the problems in practice.