CI: upload _install_ci instead of _build_ci
This means we shouldn't run dependencies of the current job, so we
stop going through Makefile.ci.

Also remove duplicated transitive deps.

Probably won't work on jobs like elpi_test which expect to run in the
build dir of its dependencies, but let's see how it does for the rest.