check all keys
currently, when the unification algorithm (evarconv) tries to solve a projection problem `proj _ = RHS`, it tries the keys it saved when unfolding the RHS and stops at the first key that has a match in the canonical structure table. Now, it stops at the first key that has a match in the canonical structure table *and* leads to a successful unification problem.