Ensure at runtime that the modpath is synced with module bodies and types.
We also seize the opportunity to remove subsequent calls to mod_mp once
we have asserted that they are equal to some other modpath. This will be
less work when actually removing the modpaths from modules for good.
Assuming no assertion fails, the new code is in particular equivalent to
the old one.