(known bugged) Fix some calls to Mod_subst.add_delta_resolver.
As revealed by the implementation, this function expects that its
first argument is a delta resolver for a more precise path than its
second argument. Intuitively, this means that the first argument
lives in a (not necessarily strict) submodule of the first argument.

Some calls to the kernel were breaking this invariant. There may be
a way to trigger nasty unsoundness issues out of this, although in all
cases the resolvers looked like they were orthogonal and thus the
operation was probably commutative.