Do not use the delta_resolver type for safe environment paramresolver.
This usage is the only one subtly breaking an implicit invariant of the
delta resolver, namely that all entries share a common non-trivial modpath
root. Instead, this datatype morally contains a set of disjoint
delta-resolvers for each library and module parameter root loaded in the
environment.

We reflect this using an ad-hoc datatype with the corresponding API and
runtime invariant checks.