Fix composition of module substitution.
This is a subtle problem that does not cause unsoundness. Rather,
it makes the composition of module substitution wasteful in its
representation.

Basically, for any module substitution σ, if (mp, δ) ∈ σ where
mp : modpath and δ : delta_resolver, then morally the only
relevant data in δ are renamings of the form (mp₀ ↦ mp₁) where
mp ⊆ mp₀ and (kn₀ ↦ kn₁) where mp ⊆ modpath(kn₀).

I say morally, as this is not literally true: there is the
corner case where mp' ∈ δ with mp' ⊆ mp but mp ∉ δ, as
find_prefix will expand any path mp' ⊆ mp ⊆ p using the data
attached to mp'. This means that there is an implicit form of
η-expansion contained in the delta resolver, as we could
replace the binding (mp', p) in δ with (mp, p ++ mp \ mp')
without changing the observable behaviour.

In any case, paths that are incomparable with mp are complete
garbage that will never be accessed.

It turns out that all calls to substitution extension respected
the invariant that their delta resolvers contained no garbage,
except Mod_subst.join. We fix this function by performing a
cleanup pass before adding the resolver.