Fix domain substitution of delta-resolvers.
The function subst_dom_codom_delta_resolver was generating nonsensical
data when applying a substitution with non-trivial delta-resolvers in its
bindings.

Basically, assuming some substitution σ it is enough to look at the
domain-codomain substitution of a singleton delta-resolver δ := [mp₀ ↦ mp₁]
to understand the issue. Assume mp₀ ↦ (mp₀', δ₀) and mp₁ ↦ (mp₁', δ₁)
are both in σ. We expect

subst_dom_codom_delta_resolver σ δ = [mp₀' ↦ mp₁'] ⊕ ε

where ε is the delta-resolver associating to subpaths of mp₀' the
corresponding value through σ in δ₁. That is, if [mp₀.p ↦ v] ∈ δ₁
we expect [mp₀'.p ↦ σ(v)] ∈ ε.

Unfortunately the current implementation generates an incorrect binding
as it adds instead [mp₀.p ↦ σ(v)] to ε. I discovered this while trying
to enforce static invariants on the delta-resolver type.

There are no files in the test-suite that trigger this behaviour, but
the stdlib does have a few instances of the issue. In the current code
this means that the delta-equivalence is incomplete as we silently lose
these additional equations. Maybe it is even unsound if we fiddle a bit
with bound modules, but I do not want to try to find an exploit. In
any case I added a minimized example to the test-suite so as to keep
track of this behaviour.