Fix composition of delta resolvers in "with Module".
The previous code was breaking the implicit invariant that the left
argument as expected to live in a subpath of the right argument. This
was behaving as expected though, because there was also an implicit
lifting operation of the resolver at the same time. In the current
implementation, lifting is just the identity, but this will change.

This commit makes the lifting explicit without changing any observable
behaviour.