Faster functor application.
Instead of repeatedly substituting the arguments in the same body, we
progpagate the substitution around. This means that for a functor
with n arguments we used to perform at least n substitutions of the body,
while now these are conflated into a single substitution.

This is basically what we do with term application already.