Keep qvar in template poly entry
This slightly changes behaviour for inductives which don't really need
template, eg `Inductive foo (A:Type@{i}) : Type@{i} := .` is now not
pseudo sort poly because the parameter and conclusion qualities end up
unrelated, so the conclusion quality has no binders and must be
collapsed to Type.

There is currently no way to explicitly specify it, as
`Inductive foo@{q|i|} ...` won't work because the named `q` can't be
made above prop (and we need it above prop for evar normalization to
substitute it away in relevances).