Rename and tweak Modops.add_module_type.
Depite the name and the comment in the mli file, rather than adding a
module type, this function is actually used to add an abstract module
parameter to the environment.

This revealed that in module checking we added the module twice, since
subtyping was already adding the argument to the environment.