Handle queries in rocqshim when possible
This allows us to handle queries before trying to find Prelude.vo,
except for -where and -config which need to initialize the Rocq
environment, and -help which is different based on the subcommand and
so needs to find the subcommand exe.

As such #20128 is not fixed for -help (but is fixed for -v).

We will need to move the specific_usage data somewhere more accessible
to the shim to be able to fix -help.