Stop relying on a hook to set the TC flag of freshly declared evars.
This is a very fragile heuristic that makes an involved implicit
dependency on δ-equivalence to decide that a constant might be a TC.
Instead of grafting the TC hook a posteriori, we change the default
value of the optional flag typeclass_candidate of Evd.new_pure_evar
and fix the callers accordingly.

Wherever it was clear that the TC status of an evar did not matter we
directly set the flag to false. This is the case e.g. when the type
is statically known not to match a TC constant or when the evar is
created from within a refine block.