Use a fast path in Unify.unify_atom.
There is no point in trying to unify two terms to find the value of
a meta if it does not appear in any of these terms. We have to be
a bit careful because the caller may make a difference between the
None and the Some Phantom case depending on the phref value, so we
pass this as an additional boolean flag.