Stop relying on firstorder context to find atomic hypotheses.
Instead we rely on an assumption-like tactic. This should be virtually
backwards compatible, since the previous code used to strongly normalize
terms to compare them. The few differences with the previous algorithm is
that it is linear in the size of the context and that it now handle η
rules as well. The first point is a regression but in a next commit we will
remove the normalization pass of all hypotheses, which should more than
compensate.