Fix guard checker "internal subterm" handling
Fix #20455

Based on analysis by Yann Leray:

>The issue that lead to this proof is that in (inductive.ml line 803-806), the
"internally bound subterms" contained in the first argument of Subterm
are forgotten instead of being preserved in the construction of the
subtree spec.

IDK how correct & complete this fix is, but it seems to work and the
code change doesn't look particularly adhoc.